pyproject.toml
[project]
name = "foglioai"
version = "0.1.0"
description = "A vintage newspaper-style article generator powered by AI agents"
authors = [
    {name = "Tommaso Elia",email = "tommaso.elia@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "redis (>=6.0.0,<7.0.0)",
    "crewai (>=0.118.0,<0.119.0)",
    "mistralai (>=1.7.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pydantic-settings (>=2.2.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "readability-lxml (>=0.8.1,<0.9.0)",
    "lxml[html-clean] (>=5.4.0,<6.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "sse-starlette (>=1.8.2,<2.0.0)",
    "pytz (>=2024.1,<2025.0)",
    "markdown (>=3.5.2,<4.0.0)"
]

[tool.poetry]
name = "foglioai"
version = "0.1.0"
description = "A vintage newspaper-style article generator powered by AI agents"
authors = ["Tommaso Elia <tommaso.elia@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
redis = "^6.0.0"
crewai = "^0.118.0"
mistralai = "^1.7.0"
beautifulsoup4 = "^4.13.4"
jinja2 = "^3.1.6"
matplotlib = "^3.10.1"
playwright = "^1.42.0"
pydantic-settings = "^2.2.1"
requests = "^2.32.3"
tenacity = "^9.1.2"
readability-lxml = "^0.8.1"
lxml = {extras = ["html-clean"], version = "^5.4.0"}
fastapi-limiter = "^0.1.6"
sse-starlette = "^1.8.2"
pytz = "^2024.1"
markdown = "^3.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
ruff = "^0.2.2"
requests-mock = "^1.12.1"
fakeredis = "^2.21.3"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
